annotate3d/demo/requirements.txt
gradio_annotate3d